More simplifications, work on RR
I have made some more simplifcations to the structure of scheduler
implementations. I have also been working on the RR scheduler. I thought
I stumbled onto a simple implementations; however, there are issues with
synchronization and what I call striping. Synchronization issues come
from the fact that different processors appear to be running different
threads of different processes as they are lagging behind a little bit -
I need to establish a mechanism by which all processors are operating on
the same process. Next this issue of striping is that instead of a
thread staying on one processor, it hops over to another. Say the
scheduler was just started, and everything is roughly in sync, running a
single thread, call it the main thread, on a process, call it "A", would
result in the following on a four processor system:

Tick | Processor 0 | Processor 1 | Processor 2 | Processor 3
0    | A.Main      | NO THREAD   | NO THREAD   | NO THREAD
1    | NO THREAD   | A.Main      | NO THREAD   | NO THREAD
2    | NO THREAD   | NO THREAD   | A.Main      | NO THREAD
3    | NO THREAD   | NO THREAD   | NO THREAD   | A.Main

In the same vein, the synchronization issue would like this using the
above representation (say there is another process, "B", also with a
single, main thread):
Tick | Processor 0 | Processor 1 | Processor 2 | Processor 3
0    | B.Main      | A.Main      | A.Main      | A.Main
1    | B.Main      | A.Main      | A.Main      | A.Main
...