RR: fix potential freeze
The tick function would switch to the next process given that the current process was not NULL. If there was no other process than the begin process, then this could cause the loss of a cycle of the scheduler. Separating the else clause into its own if statement should fix this.